requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
httpx==0.27.0
stripe==8.0.0
python-multipart==0.0.9
jinja2==3.1.4
slowapi==0.1.9
pydantic[email]==2.9.0
pytest-asyncio==0.24.0
